move some files to validate folder from utils
