+++ lib/js/_validate/_check-context-version.js
@@ -7,7 +7,7 @@
 const semver = require('semver');
 const reporter = require('@springernature/util-cli-reporter');
 
-const npmRegistryRequest = require('./_npm-registry-request');
+const npmRegistryRequest = require('../_utils/_npm-registry-request');
 
 /**
  * Find a version of context on NPM
