+++ lib/js/_validate/_check-file-naming.js
@@ -1,5 +1,5 @@
 /**
- * _get-file-naming.js
+ * _check-file-naming.js
  * Check any enforced file naming based on context dependencies
  */
 'use strict';
