+++ lib/js/_validate/_check-package-context.js
@@ -7,8 +7,8 @@
 const path = require('path');
 const reporter = require('@springernature/util-cli-reporter');
 
-const checkContextVersion = require('../_utils/_check-context-version');
-const checkFileNaming = require('../_utils/_check-file-naming');
+const checkContextVersion = require('./_check-context-version');
+const checkFileNaming = require('./_check-file-naming');
 
 /**
  * Check that brand context has been defined properly
